created stubs for preview
